requirements.txt
python-dotenv
requests
psycopg2-binary
pytest

# Airflow core package
apache-airflow==2.8.1
--constraint https://raw.githubusercontent.com/apache/airflow/constraints-2.8.1/constraints-3.11.txt
